refacto toast in useToast hook
