+++ src/domains/core/component/custom-hooks/use-toast/use-toast.scss
@@ -2,7 +2,7 @@
 	position: absolute;
 	top: 70px;
 	right: 0px;
-	width: 300px;
+	min-width: 25%;
 	font-size: 14px;
 	padding: 20px;
 	border-radius: 3px;
