+++ src/domains/core/component/top-navigation/top-navigation.js
@@ -7,15 +7,14 @@ import { AuthContext } from "../../../user/auth/auth-context.js";
 import { logout } from "../../../user/auth/auth-services.js";
 import ReminderBtn from "../../../reminder/component/reminder-btn/reminder-btn";
 import useOnClickOutside from "../custom-hooks/use-on-click-outside.js";
-import useModal from "../custom-hooks/use-modal.js";
-import Toast from "../toast/toast";
+import useToast from "../custom-hooks/use-toast/use-toast";
 import "./top-navigation.scss";
 
 const TopNavigation = ({ children }) => {
 	const { user } = useContext(AuthContext);
 	const history = useHistory();
 	const { t } = useTranslation();
-	const { ModalWrapper, open, close } = useModal();
+	const { Toast, open, close } = useToast();
 
 	const menuRef = useRef();
 	const [ showMenu, setShowMenu ] = useState(false);
@@ -26,12 +25,10 @@ const TopNavigation = ({ children }) => {
 			.then(() => history.push("/sign-in"))
 			.catch(() => open());
 	};
-	
+
 	return (
 		<header className="top-navigation">
-			<ModalWrapper>
-				<Toast autoDelete autoDeleteTime={5000} close={close}/>
-			</ModalWrapper>
+			<Toast autoDeleteTime={5000} close={close}/>
 			<Link to={"/"}>
 				<img src={logo} alt="reminder poster" className="top-navigation__logo"/>
 			</Link>
